Moved make_policy_decision() into library/policy.c
